fix bug where destructor is run without being init
There is an issue with a version of pthreads without a null check in
pthread_key_delete. Since the pthread_key has not been created yet (as
init_once hasn't been run) it is null. This logic is conditional on the
init flag being set, but the flag is set from the get-go. The fix is to
simply start with the flag toggled off as the author probably meant.